Sécurisation du CRUD des comptes utilisateur avec un voter
Filtrage des utilisateurs dans le CRUD des comptes utilisateur.
Création d'un voter pour décider des autorisations d'accès aux données.

+++ src/Controller/UserController.php
@@ -24,6 +24,9 @@ class UserController extends AbstractController
     #[Route('/', name: 'app_user_index', methods: ['GET'])]
     public function index(UserRepository $userRepository): Response
     {
+        // seul le compte admin peut accéder à la liste de tous les comptes utilisateurs
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         return $this->render('user/index.html.twig', [
             'users' => $userRepository->findAll(),
         ]);
@@ -32,6 +35,9 @@ class UserController extends AbstractController
     #[Route('/new', name: 'app_user_new', methods: ['GET', 'POST'])]
     public function new(Request $request, UserRepository $userRepository): Response
     {
+        // seul le compte admin peut créer un nouveau compte utilisateur
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         $user = new User();
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
@@ -56,6 +62,9 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_show', methods: ['GET'])]
     public function show(User $user): Response
     {
+        // seul le compte admin ou le propriétaire du compte peut afficher un compte utilisateur
+        $this->denyAccessUnlessGranted('view', $user);
+
         return $this->render('user/show.html.twig', [
             'user' => $user,
         ]);
@@ -64,6 +73,9 @@ class UserController extends AbstractController
     #[Route('/{id}/edit', name: 'app_user_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, User $user, UserRepository $userRepository): Response
     {
+        // seul le compte admin ou le propriétaire du compte peut modifier un compte utilisateur
+        $this->denyAccessUnlessGranted('edit', $user);
+
         $form = $this->createForm(UserType::class, $user);
         $form->handleRequest($request);
 
@@ -87,6 +99,9 @@ class UserController extends AbstractController
     #[Route('/{id}', name: 'app_user_delete', methods: ['POST'])]
     public function delete(Request $request, User $user, UserRepository $userRepository): Response
     {
+        // seul le compte admin peut supprimer un compte utilisateur existant
+        $this->denyAccessUnlessGranted('ROLE_ADMIN');
+
         if ($this->isCsrfTokenValid('delete'.$user->getId(), $request->request->get('_token'))) {
             $userRepository->remove($user, true);
         }
